share/legibility refactor
