+++ apps/main/app/features/scenarioExplorer/explorer/tools/panels/resilience/resilienceHydroclimates.ts
@@ -1,4 +1,15 @@
-/** Hydroclimate keys used by the resilience heatmap matrix and store defaults */
-export const RESILIENCE_HYDROCLIMATES = ["historical", "cc50", "cc95"] as const
+import {
+  HYDROCLIMATES,
+  type Hydroclimate,
+} from "../../../../../../content/scenarios"
 
-export type ResilienceHydroclimate = (typeof RESILIENCE_HYDROCLIMATES)[number]
+/**
+ * Hydroclimates used by the resilience heatmap matrix and store
+ * defaults. These are the app-wide hydroclimates (HYDROCLIMATES in
+ * content/scenarios). The resilience-specific aliases stay so the
+ * resilience code that imports them reads clearly, and so this module
+ * remains hook-free for the store-init path.
+ */
+export const RESILIENCE_HYDROCLIMATES = HYDROCLIMATES
+
+export type ResilienceHydroclimate = Hydroclimate
